requirement.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
streamlit-autorefresh==0.0.4
paho-mqtt==1.6.1